refactor(mail): refactor mail service & move `mailersend` into plugin
