+++ pyproject.toml
@@ -20,7 +20,6 @@ dependencies = [
     "httpx>=0.28.1",
     "jinja2>=3.1.0",
     "loguru>=0.7.3",
-    "mailersend>=2.0.0",
     "maxminddb>=2.8.2",
     "newrelic>=10.1.0",
     "passlib[bcrypt]>=1.7.4",
