+++ static/templates/email/password_reset_en.html
@@ -1,0 +1,140 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Password Reset - {{ server_name }}</title>
+    <style>
+        body {
+            margin: 0;
+            padding: 0;
+            font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Arial, sans-serif;
+            background-color: #f5f5f5;
+        }
+        .container {
+            max-width: 600px;
+            margin: 0 auto;
+            background-color: #ffffff;
+        }
+        .header {
+            background: linear-gradient(135deg, #ED8EA6 0%, #FF6B9D 100%);
+            color: white;
+            padding: 40px 20px;
+            text-align: center;
+        }
+        .header h1 {
+            margin: 0;
+            font-size: 28px;
+            font-weight: 600;
+        }
+        .header p {
+            margin: 10px 0 0 0;
+            font-size: 16px;
+            opacity: 0.9;
+        }
+        .content {
+            padding: 40px 30px;
+        }
+        .greeting {
+            font-size: 18px;
+            font-weight: 600;
+            color: #333;
+            margin-bottom: 20px;
+        }
+        .message {
+            font-size: 15px;
+            line-height: 1.6;
+            color: #555;
+            margin-bottom: 30px;
+        }
+        .code-box {
+            background: #f8f9fa;
+            border: 2px solid #ED8EA6;
+            border-radius: 12px;
+            padding: 25px;
+            text-align: center;
+            margin: 30px 0;
+        }
+        .code {
+            font-size: 36px;
+            font-weight: bold;
+            letter-spacing: 8px;
+            color: #ED8EA6;
+            font-family: 'Courier New', monospace;
+        }
+        .expiry {
+            font-size: 14px;
+            color: #666;
+            margin-top: 15px;
+        }
+        .danger-box {
+            background: #f8d7da;
+            border-left: 4px solid #dc3545;
+            border-radius: 6px;
+            padding: 20px;
+            margin: 30px 0;
+        }
+        .danger-box h3 {
+            margin: 0 0 10px 0;
+            font-size: 16px;
+            color: #721c24;
+        }
+        .danger-box ul {
+            margin: 10px 0;
+            padding-left: 20px;
+            color: #721c24;
+        }
+        .danger-box li {
+            margin: 8px 0;
+            font-size: 14px;
+        }
+        .footer {
+            background: #2c3e50;
+            color: #ecf0f1;
+            padding: 30px;
+            text-align: center;
+            font-size: 13px;
+        }
+        .footer p {
+            margin: 5px 0;
+            line-height: 1.5;
+        }
+    </style>
+</head>
+<body>
+    <div class="container">
+        <div class="header">
+            <h1>Password Reset</h1>
+            <p>Password Reset Request</p>
+        </div>
+
+        <div class="content">
+            <div class="greeting">Hello, {{ username }}!</div>
+
+            <div class="message">
+                <p>We received a request to reset your password. If this was you, please use the following verification code to reset your password:</p>
+            </div>
+
+            <div class="code-box">
+                <div class="code">{{ code }}</div>
+                <div class="expiry">⏱️ This code will expire in <strong>{{ expiry_minutes }} minutes</strong></div>
+            </div>
+
+            <div class="danger-box">
+                <h3>⚠️ Security Notice</h3>
+                <ul>
+                    <li>Do not share this verification code with anyone</li>
+                    <li>If you did not request a password reset, please ignore this email immediately</li>
+                    <li>This code can only be used once</li>
+                    <li>We recommend setting a strong password to protect your account</li>
+                </ul>
+            </div>
+        </div>
+
+        <div class="footer">
+            <p>© {{ year }} {{ server_name }}. All rights reserved.</p>
+            <p>This email was sent automatically, please do not reply.</p>
+        </div>
+    </div>
+</body>
+</html>
