+++ static/templates/email/password_reset_en.txt
@@ -1,0 +1,17 @@
+Hello, {{ username }}!
+
+We received a request to reset your password. If this was you, please use the following verification code to reset your password:
+
+Verification Code: {{ code }}
+
+This code will expire in {{ expiry_minutes }} minutes.
+
+Security Notice:
+- Do not share this verification code with anyone
+- If you did not request a password reset, please ignore this email immediately
+- This code can only be used once
+- We recommend setting a strong password to protect your account
+
+---
+© {{ year }} {{ server_name }}. All rights reserved.
+This email was sent automatically, please do not reply.
