+++ static/templates/email/password_reset_zh.txt
@@ -1,0 +1,17 @@
+你好，{{ username }}！
+
+我们收到了您的密码重置请求。如果这是您本人操作，请使用以下验证码重置密码：
+
+验证码：{{ code }}
+
+验证码将在 {{ expiry_minutes }} 分钟内有效。
+
+安全提醒：
+- 请不要与任何人分享这个验证码
+- 如果您没有请求密码重置，请立即忽略这封邮件
+- 验证码仅能使用一次
+- 建议设置一个强密码以保护您的账户安全
+
+---
+© {{ year }} {{ server_name }}. 保留所有权利。
+此邮件由系统自动发送，请勿回复。
