+++ static/templates/email/verification_en.txt
