+++ static/templates/email/verification_zh.txt
